moved all functions to searchbar and onsubmit takes user to dashboard with search params and removed unused code
